+++ Assets/Scripts/Composite/FinalBonus.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 public class FinalBonus : BaseAttribute {
 
-    public FinalBonus(int value = 0, int multiplier = 0) : base(value, multiplier) {
+    public FinalBonus(int value = 0, float multiplier = 0) : base(value, multiplier) {
 
     }
